pyproject.toml
[tool.poetry]
name = "creative-tools"
version = "0.1.0"
description = ""
authors = ["dnl-strnl <129630837+dnl-strnl@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.11"
clip-interrogator = "^0.6.0"
datasets = "^2.17.0"
hydra-core = "^1.3.2"
milvus = "^2.3.5"
numpy = "^1.26.4"
omegaconf = "^2.3.0"
pillow = "^10.2.0"
torchvision = "^0.17.0"
tqdm = "^4.66.2"
pycocotools = "^2.0.7"
opencv-python = "^4.9.0.80"
matplotlib = "^3.8.3"
timm = "^0.9.16"
transformers = "^4.37.2"
mlx-lm = "^0.0.13"
mlx = "^0.4.0"
dvc = "^3.30.0"
dvclive = "^3.30.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[[tool.poetry.packages]]
include = "creative_tools"
from = "src"
